Added tests for line_plot
